+++ tests/unit/crystallography/Lattice_test.cpp
@@ -145,24 +145,16 @@ TEST(LatticeTest, JsonRoundtripTest) {
                              Lattice::hexagonal()}) {
     // --- to_dict equivalent ---
     jsonParser json_out;
-    std::cout << "here 1" << std::endl;
     json_out["lattice_vectors"] = lat.lat_column_mat().transpose();
     // upcast to nlohmann::json as the binding does
-    std::cout << "here 2" << std::endl;
-    nlohmann::json data = json_out;
+    nlohmann::json const data = json_out;
 
     // --- from_dict equivalent ---
-    std::cout << "here 3" << std::endl;
     jsonParser json{data};
-    std::cout << "here 4" << std::endl;
     Eigen::Matrix3d latvec_transpose;
-    std::cout << "here 5" << std::endl;
     ASSERT_NO_THROW(from_json(latvec_transpose, json["lattice_vectors"]));
-    std::cout << "here 6" << std::endl;
     Lattice roundtrip(latvec_transpose.transpose(), tol);
 
-    std::cout << "here 7" << std::endl;
     EXPECT_TRUE(xtal::is_equivalent(lat, roundtrip));
-    std::cout << "here 8" << std::endl;
   }
 }